Adding Agency Repository to Agency Client Repository
Using the composition of one aggregate to validate on another aggregate.
Each applies the invariance specific to it. Since deleting a client
would not affect the consultant roles they are not part of the
aggregate.

Agency Client Consultant Roles do not have a client_id and as such are
not part of that aggregate.

Maybe a boundary here that is not defined correctly atm?