POC - Read Projections
POC of the read projections.
This covers the UI Management Screens

This is not the final approach though and there are a couple of real
considerations that needs to be applied here before we decide on the
final implementation details.

Also the token writing was not implemented here as this just means more
refctor when converting this to typescript